+++ verifier/src/lib.rs
@@ -7,7 +7,7 @@ extern crate std;
 
 use alloc::vec::Vec;
 
-use miden_air::{ProcessorAir, PublicInputs, config};
+use miden_air::{PublicInputs, config};
 use miden_crypto::stark;
 
 // RE-EXPORTS
@@ -132,7 +132,8 @@ fn verify_stark(
     let pub_inputs =
         PublicInputs::new(program_info, stack_inputs, stack_outputs, pc_transcript_state);
     let public_values = pub_inputs.to_elements();
-    let air = ProcessorAir::new();
+
+    let air = miden_air::ProcessorAir::new();
 
     match hash_fn {
         HashFunction::Blake3_256 => {